Bug fix
We had an issue with one of our customers Odoo getting an "Uncaught TypeError: Cannot read property '0' of undefined" error. I traced it back to this section and noticed that you were checking if item[group_bys[0]] was defined but then assigning item[group_bys][0]. I fixed this and the line after. After this the error went away.